eliminated all compilers warnings.
svn path=/trunk/; revision=20975


Former-commit-id: 2028e22b0984292bde0579bd542f9e89eaa6e97f

+++ v3/common/amx-api/src/main/java/com/sun/appserv/management/util/jmx/MBeanServerConnection_Perf.java
@@ -133,7 +133,7 @@ public class MBeanServerConnection_Perf
 		printTime( final long id, final String methodName, final Object[] args )
 		{
 			final long	curTime	= System.currentTimeMillis();
-			final Long	start	= (Long)mTimers.remove( new Long( id ) );
+			final Long	start	= mTimers.remove( new Long( id ) );
 			if ( start != null )
 			{
 				mOutput.println( getInvocationString( methodName, args ) + ": " +
